se-03-team-29-main/requirements.txt
cffi==1.15.0
click==7.1.2
colorama==0.4.4
cryptography==36.0.2
Flask==1.1.4
Flask-MySQLdb==1.0.1
flask-selfdoc==1.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
Pillow==9.1.0
protobuf==3.20.1
pycparser==2.21
pyOpenSSL==22.0.0
PyYAML==6.0
qrcode==7.3.1
urllib3==1.26.9
Werkzeug==1.0.1
